test: extract svar_link/Metadata pydantic unit tests to unit/dataset/
