+++ tests/unit/dataset/test_svar_link_models.py
@@ -1,0 +1,95 @@
+"""Unit tests for ``SvarLink`` / ``Metadata`` Pydantic models and pure
+``_resolve_svar`` error-path logic.
+
+Originally lived in ``tests/integration/dataset/test_svar_link.py``;
+extracted to the unit tier because each test here either exercises only
+Pydantic model construction/validation/serialization or uses ``tmp_path``
+to drive an error path in ``_resolve_svar``. None requires a written GVL
+dataset.
+
+The 14 integration-tier tests in the original file all take the
+``svar_dataset_paths`` fixture (which calls ``gvl.write``) or otherwise
+require a real written dataset — they remain in the integration tier.
+"""
+
+import json
+
+import pytest
+from pydantic import ValidationError
+from pydantic_extra_types.semantic_version import SemanticVersion
+
+from genvarloader._dataset._svar_link import (
+    SvarFingerprint,
+    SvarLink,
+    _resolve_svar,
+)
+from genvarloader._dataset._write import Metadata
+
+
+def test_svar_link_roundtrip():
+    link = SvarLink(
+        relative_path="../foo.svar",
+        absolute_path="/abs/path/foo.svar",
+        fingerprint=SvarFingerprint(n_variants=10, variant_idxs_bytes=42),
+    )
+    payload = link.model_dump_json()
+    parsed = SvarLink.model_validate_json(payload)
+    assert parsed == link
+
+
+def test_svar_link_rejects_malformed_fingerprint():
+    bad = (
+        '{"relative_path":"a","absolute_path":"b",'
+        '"fingerprint":{"n_variants":"not_an_int","variant_idxs_bytes":1}}'
+    )
+    with pytest.raises(ValidationError):
+        SvarLink.model_validate_json(bad)
+
+
+def test_metadata_version_parses_existing_strings():
+    payload = json.dumps(
+        {
+            "samples": ["s1"],
+            "contigs": ["1"],
+            "n_regions": 1,
+            "version": "0.18.0",
+        }
+    )
+    m = Metadata.model_validate_json(payload)
+    assert isinstance(m.version, SemanticVersion)
+    assert m.version == SemanticVersion.parse("0.18.0")
+
+
+def test_metadata_version_serializes_back_to_string():
+    m = Metadata(
+        samples=["s1"],
+        contigs=["1"],
+        n_regions=1,
+        version=SemanticVersion.parse("0.18.0"),
+    )
+    dumped = json.loads(m.model_dump_json())
+    assert dumped["version"] == "0.18.0"
+
+
+def test_metadata_svar_link_defaults_to_none():
+    m = Metadata(samples=["s1"], contigs=["1"], n_regions=1)
+    assert m.svar_link is None
+
+
+def test_semantic_version_ordering_for_one_based_dispatch():
+    """The legacy comparison '>= 0.18.0' must still work under SemanticVersion."""
+    assert SemanticVersion.parse("0.18.0") >= SemanticVersion.parse("0.18.0")
+    assert SemanticVersion.parse("0.20.0") >= SemanticVersion.parse("0.18.0")
+    assert not (SemanticVersion.parse("0.17.5") >= SemanticVersion.parse("0.18.0"))
+
+
+def test_resolve_svar_raises_when_not_found(tmp_path):
+    gvl_path = tmp_path / "ds.gvl"
+    gvl_path.mkdir()
+    link = SvarLink(
+        relative_path="nowhere",
+        absolute_path="/nowhere",
+        fingerprint=SvarFingerprint(n_variants=1, variant_idxs_bytes=1),
+    )
+    with pytest.raises(FileNotFoundError, match="svar="):
+        _resolve_svar(gvl_path, link, override=None)
